requirements.txt
python-telegram-bot>=21.0
openpyxl>=3.1.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
